fix(auth)!: fix some serious bugs related to authentication (IDOR), optimization. Also slightly refactor code, change structure. Update tests to make them follow updated logic.

+++ src/main/java/team/anonyms/converter/controllers/frontend/ModificationController.java
@@ -47,7 +47,7 @@ public class ModificationController {
     }
 
     @GetMapping("/{patternId}")
-    public ResponseEntity<Integer> getNumberOfAllModificationsByPatternId(
+    public ResponseEntity<Long> getNumberOfAllModificationsByPatternId(
             @PathVariable UUID patternId
     ) {
         logger.info("Called getNumberOfModificationsByPatternId; patternId={}", patternId);
